requirements.txt
Flask
gunicorn
scikit-learn
pandas
matplotlib
numpy
PyYAML
-e .